# feat: Human-readable exception names and 25x performance improvement