… CLAUDE.md
… tests/emitters/test_exceptions_emitter.py
… tests/visit/endpoint/generators/test_response_handler_generator.py
… tests/visit/endpoint/generators/test_response_handler_generator_strategy.py